Versión Final 1.0
Eliminación de la mala implementación del patrón State. Corrección en la justificación, diagrama, etc.

Implementación de clases Estado y Posible Estado.

+++ src/main/java/extras/modos_de_viaje/DetenerseNMinutos.java
@@ -21,7 +21,7 @@ public class DetenerseNMinutos implements ModoDeViaje{
         int demoraPorDetenciones = (trayecto.getTramos().size()-1)*minutosDetenido;
 
         //Demora sólo de los trayectos, sin contar detenciones
-        float demoraTotalTrayecto = 0;
+        double demoraTotalTrayecto = 0;
         for(Tramo unTramo : trayecto.getTramos()){
             demoraTotalTrayecto+= unTramo.calcularDemoraTramo();
         }
